add rocket on mobile
