+++ components/Section.vue
@@ -6,7 +6,7 @@
     .section__header(v-else)
       h2.section__title
         a(:href="'#' + slug(title)") {{title}}
-      button.section__rocket(@click="goToTop()" ref='rocket')
+      Rocket
     slot
     .section__intersection.section__intersection--bottom
 </template>
@@ -20,19 +20,6 @@ export default Vue.extend({
   methods: {
     slug(str: string) {
       return slugify(str, { lower: true });
-    },
-    goToTop() {
-      const rocket = this.$refs.rocket as HTMLElement;
-      rocket.style.transform = 'translate(20px, -200%) rotate(-45deg)';
-
-      window.scrollTo({ behavior: 'smooth', top: 0, left: 0 });
-      const headerTitleAnchor = document.querySelector(
-        '.header__title a'
-      ) as HTMLAnchorElement;
-      setTimeout(() => {
-        headerTitleAnchor.focus();
-        rocket.style.transform = '';
-      }, 1000);
     },
   },
 });
@@ -100,56 +87,6 @@ export default Vue.extend({
       color: color-get('white');
       text-decoration: none;
       @extend %text-shadow;
-    }
-  }
-
-  @include e('rocket') {
-    width: $text-30-lineheight + $padding * 2;
-    height: $text-30-lineheight + $padding * 2;
-    -webkit-appearance: none;
-    background-color: transparent;
-    border: 0;
-    cursor: pointer;
-    padding: 0;
-    transform: translate($padding, -$padding) rotate(0deg);
-    transition: transform linear 200ms;
-    background-size: $text-30-lineheight;
-    position: relative;
-
-    @include bp('grid-bigger') {
-      width: $text-40-lineheight + $padding * 2;
-      height: $text-40-lineheight + $padding * 2;
-    }
-
-    body.touch & {
-      display: none;
-    }
-
-    &:hover,
-    &:focus-visible {
-      transform: translate($padding, -$padding) rotate(-45deg);
-
-      &::before {
-        animation: anim-drop-shadow-hover $shadow-hover-anim;
-      }
-    }
-
-    &::before {
-      content: '';
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      background-position: center;
-      background-size: $text-30-lineheight;
-      background-repeat: no-repeat;
-      background-image: url('~/assets/icons/rocket.svg');
-      filter: drop-shadow(10px -5px 10px black);
-
-      @include bp('grid-bigger') {
-        background-size: $text-40-lineheight;
-      }
     }
   }
 }
